change portfolio open format version to 1.1.0
