+++ src/main/java/ru/investbook/openformat/PortfolioOpenFormatRestController.java
@@ -28,9 +28,9 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
-import ru.investbook.openformat.v1_0_0.PortfolioOpenFormatBuilder;
-import ru.investbook.openformat.v1_0_0.PortfolioOpenFormatPersister;
-import ru.investbook.openformat.v1_0_0.PortfolioOpenFormatV1_0_0;
+import ru.investbook.openformat.v1_1_0.PortfolioOpenFormatBuilder;
+import ru.investbook.openformat.v1_1_0.PortfolioOpenFormatPersister;
+import ru.investbook.openformat.v1_1_0.PortfolioOpenFormatV1_1_0;
 
 import javax.servlet.http.HttpServletResponse;
 import java.io.ByteArrayOutputStream;
@@ -57,7 +57,7 @@ public class PortfolioOpenFormatRestController {
         try {
             long t0 = System.nanoTime();
             String fileName = "portfolio.json";
-            PortfolioOpenFormatV1_0_0 object = portfolioOpenFormatFactory.create();
+            PortfolioOpenFormatV1_1_0 object = portfolioOpenFormatFactory.create();
             sendSuccessHeader(response, fileName, "application/json");
             ByteArrayOutputStream outputStream = new ByteArrayOutputStream(10240);
             objectMapper.writerWithDefaultPrettyPrinter().writeValue(outputStream, object);
@@ -74,7 +74,7 @@ public class PortfolioOpenFormatRestController {
     @PostMapping("upload")
     public ResponseEntity<String> upload(@RequestParam("file") MultipartFile file) {
         try (InputStream inputStream = file.getInputStream()) { // creates new input stream
-            PortfolioOpenFormatV1_0_0 object = objectMapper.readValue(inputStream, PortfolioOpenFormatV1_0_0.class);
+            PortfolioOpenFormatV1_1_0 object = objectMapper.readValue(inputStream, PortfolioOpenFormatV1_1_0.class);
             portfolioOpenFormatPersister.persist(object);
             return ok();
         } catch (Exception e) {
